# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Appointment Booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  synced: "synced",
} as const;

export interface Appointment {
  id: number;
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  status: AppointmentStatus;
  externalId?: string | null;
  source?: string | null;
  clinicId?: number | null;
  clinicName?: string | null;
  createdAt: string;
}

export interface CreateAppointmentRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  phone: string;
  date: string;
  time: string;
  clinicId?: number | null;
}

export interface Clinic {
  id: number;
  name: string;
}

export interface TimeSlot {
  time: string;
  available: boolean;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListAppointmentsParams = {
  clinicId?: number;
  status?: string;
};

export type GetAvailableSlotsParams = {
  date: string;
  clinicId?: number;
};