# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Medical Appointment Booking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: appointments
    description: Appointment operations
  - name: clinics
    description: Clinic operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: clinicId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get a single appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      summary: List clinics
      responses:
        "200":
          description: List of clinics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinic"

  /available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: clinicId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Appointment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, synced]
        externalId:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        clinicId:
          type: integer
          nullable: true
        clinicName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - date
        - time
        - status
        - createdAt

    CreateAppointmentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 6
        date:
          type: string
          format: date
        time:
          type: string
        clinicId:
          type: integer
          nullable: true
      required:
        - name
        - email
        - phone
        - date
        - time

    Clinic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name

    TimeSlot:
      type: object
      properties:
        time:
          type: string
        available:
          type: boolean
      required:
        - time
        - available

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message