# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Appointment Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface Appointment {
  id: number;
  name: string;
  email: string;
  phone: string;
  date: Date;
  time: string;
  status: AppointmentStatus;
  externalId?: string | null;
  source?: string | null;
  clinicId?: number | null;
  clinicName?: string | null;
  createdAt: Date;
}