# lib/api-zod/src/generated/types/createAppointmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Appointment Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  phone: string;
  date: Date;
  time: string;
  clinicId?: number | null;
}