# lib/api-zod/src/generated/types/timeSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Appointment Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeSlot {
  time: string;
  available: boolean;
}